Rename webview-interface file to webview-definition
