+++ extensions/ql-vscode/src/view/compare/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { WebviewDefinition } from '../webview-interface';
+import { WebviewDefinition } from '../webview-definition';
 import { Compare } from './Compare';
 
 const definition: WebviewDefinition = {
