+++ extensions/ql-vscode/src/view/remote-queries/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { WebviewDefinition } from '../webview-interface';
+import { WebviewDefinition } from '../webview-definition';
 import { RemoteQueries } from './RemoteQueries';
 
 const definition: WebviewDefinition = {
