+++ extensions/ql-vscode/src/view/results/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { WebviewDefinition } from '../webview-interface';
+import { WebviewDefinition } from '../webview-definition';
 import { ResultsApp } from './results';
 
 const definition: WebviewDefinition = {
