+++ extensions/ql-vscode/src/view/variant-analysis/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { WebviewDefinition } from '../webview-interface';
+import { WebviewDefinition } from '../webview-definition';
 import { VariantAnalysis } from './VariantAnalysis';
 
 const definition: WebviewDefinition = {
